Byron: move code from tests to test lib
So we can reuse it for the Cardano tests.
