+++ ouroboros-consensus-byron-test/src/Test/ThreadNet/Infra/Byron/Genesis.hs
@@ -1,0 +1,74 @@
+{-# LANGUAGE NamedFieldPuns    #-}
+{-# LANGUAGE OverloadedStrings #-}
+module Test.ThreadNet.Infra.Byron.Genesis (
+    realPBftParams
+  , generateGenesisConfig
+  ) where
+
+import           Control.Monad.Except (runExceptT)
+import           Data.Time (Day (..), UTCTime (..))
+
+import           Ouroboros.Consensus.BlockchainTime
+import           Ouroboros.Consensus.Config.SecurityParam
+import           Ouroboros.Consensus.Node.ProtocolInfo
+import           Ouroboros.Consensus.Protocol.PBFT
+
+import qualified Cardano.Chain.Common as Common
+import qualified Cardano.Chain.Genesis as Genesis
+import qualified Cardano.Chain.Update as Update
+import qualified Cardano.Crypto as Crypto
+import qualified Test.Cardano.Chain.Genesis.Dummy as Dummy
+
+import           Ouroboros.Consensus.Byron.Ledger.Conversions
+
+{-------------------------------------------------------------------------------
+  Generating the genesis configuration
+-------------------------------------------------------------------------------}
+
+realPBftParams :: SecurityParam -> NumCoreNodes -> PBftParams
+realPBftParams paramK numCoreNodes = PBftParams
+  { pbftNumNodes           = numCoreNodes
+  , pbftSecurityParam      = paramK
+  , pbftSignatureThreshold = (1 / n) + (1 / k) + epsilon
+    -- crucially: @floor (k * t) >= ceil (k / n)@
+  }
+    where
+      epsilon = 1/10000   -- avoid problematic floating point round-off
+
+      n :: Num a => a
+      n = fromIntegral x where NumCoreNodes x = numCoreNodes
+
+      k :: Num a => a
+      k = fromIntegral x where SecurityParam x = paramK
+
+-- Instead of using 'Dummy.dummyConfig', which hard codes the number of rich
+-- men (= CoreNodes for us) to 4, we generate a dummy config with the given
+-- number of rich men.
+generateGenesisConfig :: SlotLength
+                      -> PBftParams
+                      -> (Genesis.Config, Genesis.GeneratedSecrets)
+generateGenesisConfig slotLen params =
+    either (error . show) id $
+      Crypto.deterministic "this is fake entropy for testing" $
+        runExceptT $
+          Genesis.generateGenesisConfigWithEntropy startTime spec
+  where
+    startTime = UTCTime (ModifiedJulianDay 0) 0
+    PBftParams{pbftNumNodes, pbftSecurityParam} = params
+    NumCoreNodes numCoreNodes = pbftNumNodes
+
+    spec :: Genesis.GenesisSpec
+    spec = Dummy.dummyGenesisSpec
+        { Genesis.gsInitializer = Dummy.dummyGenesisInitializer
+          { Genesis.giTestBalance =
+              (Genesis.giTestBalance Dummy.dummyGenesisInitializer)
+                -- The nodes are the richmen
+                { Genesis.tboRichmen = fromIntegral numCoreNodes }
+          }
+        , Genesis.gsK = Common.BlockCount $ maxRollbacks pbftSecurityParam
+        , Genesis.gsProtocolParameters = gsProtocolParameters
+          { Update.ppSlotDuration = toByronSlotLength slotLen
+          }
+        }
+      where
+        gsProtocolParameters = Genesis.gsProtocolParameters Dummy.dummyGenesisSpec
